+++ activerecord/lib/active_record/connection_adapters/postgresql_adapter.rb
@@ -166,6 +166,10 @@ module ActiveRecord
         true
       end
 
+      def supports_check_constraints?
+        true
+      end
+
       def supports_validate_constraints?
         true
       end